Allow overriding of the service IP
Currently the service IP is always assigned based on the bind_address
setting. For the api-int service we need to be able to specify a
different IP (the API VIP). This change adds an optional "ip" parameter
to the service config which, if set, overrides bind_address as the
AddrIPv4 value for the service.