Exit if interface loses IP
On startup, mdns-publisher looks up which interface has the configured
IP. However, it is possible for this IP to be subsequently moved,
such as when the interface gets added to a bridge. In that case, we
need to be restarted and re-detect the bridge or our records become
inaccessible.

This change adds a goroutine that loops every five seconds checking
if the detected interface still has the configured IP. If it does not,
it closes a channel to indicate to the main thread that it should exit.
This should allow kubernetes to restart it on the correct interface.